+++ pkg/sql/compile/operator.go
@@ -886,7 +886,7 @@ func constructWindow(ctx context.Context, n *plan.Node, proc *process.Process) *
 	for i, expr := range n.WinSpecList {
 		f := expr.Expr.(*plan.Expr_W).W.WindowFunc.Expr.(*plan.Expr_F)
 		distinct := (uint64(f.F.Func.Obj) & function.Distinct) != 0
-		obj := int(uint64(f.F.Func.Obj) & function.DistinctMask)
+		obj := int64(uint64(f.F.Func.Obj) & function.DistinctMask)
 		var e *plan.Expr = nil
 		var cfg []byte
 
@@ -959,7 +959,7 @@ func constructGroup(ctx context.Context, n, cn *plan.Node, ibucket, nbucket int,
 	for i, expr := range n.AggList {
 		if f, ok := expr.Expr.(*plan.Expr_F); ok {
 			distinct := (uint64(f.F.Func.Obj) & function.Distinct) != 0
-			obj := int(uint64(f.F.Func.Obj) & function.DistinctMask)
+			obj := int64(uint64(f.F.Func.Obj) & function.DistinctMask)
 			if len(f.F.Args) > 0 {
 				//for group concat, the last arg is separator string
 				if f.F.Func.ObjName == plan2.NameGroupConcat && len(f.F.Args) > 1 {
